move _iter_attributes to scrape function
