+++ otscrape/base/__init__.py
@@ -1,0 +1,2 @@
+from .data_model import DataModel
+from .page import Page
